Move timeline types/config into the feature folder; code-split routes
types/timeline.ts and config/timeline.ts were timeline-only files in
top-level dirs. Routes are now React.lazy so the syntax highlighter
loads only on /blog, and Markdown uses the hljs Light build with just
typescript registered: Blog chunk 934 kB -> 67 kB, initial JS 1.59 MB
-> ~490 kB, chunk-size warning gone.
