+++ src/components/timeline/AnimatedParticleContainer.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Box } from '@chakra-ui/react';
 import ParticleEffect from 'components/timeline/ParticleEffect';
-import { Particle, ParticleConfig } from 'types/timeline';
-import { timelineColors } from 'config/timeline';
+import { Particle, ParticleConfig } from 'components/timeline/types';
+import { timelineColors } from 'components/timeline/config';
 
 interface AnimatedParticleContainerProps {
   particles: Particle[];
