+++ src/components/timeline/TimelineItem.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Box, Flex } from '@chakra-ui/react';
 import { motion } from 'framer-motion';
-import { TimelineItem as TimelineItemType } from 'types/timeline';
-import { particleConfig, timelineColors } from 'config/timeline';
+import { TimelineItem as TimelineItemType } from 'components/timeline/types';
+import { particleConfig, timelineColors } from 'components/timeline/config';
 import TimelineCard from 'components/timeline/TimelineCard';
 import TimelineIcon from 'components/timeline/TimelineIcon';
 import AnimatedParticleContainer from 'components/timeline/AnimatedParticleContainer';
